docs(readme,include): fix documentation for readme and include files
Readme and include files are now properly referenced inside doxygen documentation. The README has
now a proper header and footer, the include files have been added to the correct doxygen module and
the sobel benchmark had its documentation fixed.

BREAKING CHANGE: All the other benchmarks in the src folder contain an outdate copy of the sobel
header. This  has to be fixed since all benchmarks are picked up by doxygen as copies of the sobel
benchmark

Signed-off-by: Mattia Nicolella <mattianicolella@gmail.com>